requirements.txt
aiogram
httpx
python-dotenv
fastapi
uvicorn
redis
pydantic